FslBase: fixed the issues found by the special member, enum size and empty catch checks
- cppcoreguidelines-special-member-functions:
  - The span, string view and handle vector types no longer declare copy and move members that are
    identical to the implicit ones (rule of zero).
  - The classes with only static members delete the default constructor instead of the copy
    operations.
  - The non copyable classes declare the destructor (Noncopyable makes it protected).
  - The scoped state helpers delete the copy operations.
- performance-enum-size: the enums use the smallest underlying type. None of them are serialized,
  and the stream operators cast them to uint32_t before printing.
- bugprone-empty-catch: the logging code intentionally swallows exceptions, so the catch
  statements are marked with NOLINT.
- bugprone-implicit-widening-of-multiplication-result: fixed the unit test multiplication.

+++ DemoFramework/FslBase/include/FslBase/Math/RectangleSizeRestrictionFlag.hpp
@@ -35,7 +35,7 @@
 
 namespace Fsl
 {
-  enum class RectangleSizeRestrictionFlag : uint32_t
+  enum class RectangleSizeRestrictionFlag : uint8_t
   {
     NoRestrictions = 0x00,
     Power2 = 0x01,
